// frontend/src/types/generated/BlogValidationError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BlogValidationError = "EmptyTitle" | "TitleTooLong" | "EmptyTitleRo" | "TitleRoTooLong" | "EmptySlug" | "SlugTooLong" | "InvalidSlug" | "EmptyContent" | "ContentTooLarge" | "EmptyContentRo" | "ContentRoTooLarge" | "EmptyExcerpt" | "ExcerptTooLong" | "EmptyExcerptRo" | "ExcerptRoTooLong" | "EmptyAuthor" | "AuthorTooLong";